src/harmful_meme/requirements.txt
pandas
numpy
scikit-learn
pillow  # Image Processing
transformers # Offline VLM
torch        # Offline VLM
accelerate   # Optimization
bitsandbytes # Quantization
requests     # Ollama API
